requirements.txt
discord.py
pyowm
aiohttp
websockets
chardet